requirements.txt
pip
setuptools
wheel
flask
praw
textblob
nltk
numpy<2
thinc
transformers
datasets
accelerate
elasticsearch==8.9.0
python-dotenv
hf_transfer
scikit-learn
tensorboard
# pip install torch==1.13.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html
schedule
pyyaml